+++ tests/unit/test_absolute_path.py
@@ -2,6 +2,8 @@
 
 from pathlib import Path
 
+import pytest
+
 from treestamps.tree import Treestamps
 from treestamps.tree.config import TreestampsConfig
 
@@ -35,8 +37,11 @@ class TestAbsolutePath:
         config = TreestampsConfig(program_name="test", path=root, verbose=2)
         treestamps = Treestamps(config=config)
         test_path = root
-        abs_path = treestamps._get_absolute_path(test_path, "/tmp")  # noqa: S108
-        assert abs_path is None
+        try:
+            treestamps._get_absolute_path(test_path, "/tmp")  # noqa: S108
+            assert pytest.fail("should throw an exception")
+        except ValueError:
+            assert pytest.raises(ValueError, match="outside")
 
     def test_absolute_path_subpath_parent(self) -> None:
         """Test absolute paths."""